+++ model/cce.py
@@ -43,10 +43,10 @@ class CCE(Core, Base):
 
     volume_id = Column(Integer, ForeignKey('volume.id', ondelete="CASCADE"))
 
-    registrations = relationship('Registration', backref='cce', cascade='all, delete', passive_deletes=True)
-    lccns = relationship('LCCN', backref='cce', cascade='all, delete', passive_deletes=True)
-    authors = relationship('Author', backref='cce', cascade='all, delete', passive_deletes=True)
-    publishers = relationship('Publisher', backref='cce', cascade='all, delete', passive_deletes=True)
+    registrations = relationship('Registration', backref='cce', cascade='all, delete-orphan', passive_deletes=True)
+    lccns = relationship('LCCN', backref='cce', cascade='all, delete-orphan', passive_deletes=True)
+    authors = relationship('Author', backref='cce', cascade='all, delete-orphan', passive_deletes=True)
+    publishers = relationship('Publisher', backref='cce', cascade='all, delete-orphan', passive_deletes=True)
 
     def __repr__(self):
         return '<CCE(regnums={}, uuid={}, title={})>'.format(self.registrations, self.uuid, self.title)